requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
argcomplete==2.0.0
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
frozenlist==1.3.3
idna==3.4
joblib==1.2.0
multidict==6.0.4
nltk==3.8.1
numpy==1.25.0
openai==0.27.8
pandas==2.0.2
pipx==1.0.0
pyaml==23.5.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
threadpoolctl==3.1.0
tqdm==4.65.0
tzdata==2023.3
urllib3==2.0.3
userpath==1.8.0
yarl==1.9.2
